Added firebase, maps and layout changes

+++ lib/models/monument_location.dart
@@ -5,11 +5,15 @@ class MonumentLocation {
   final double lng;
   final String zipCode;
 
-  MonumentLocation({@required this.lat, @required this.lng, @required this.zipCode})
+  MonumentLocation(
+      {@required this.lat, @required this.lng, @required this.zipCode})
       : assert(lat != null),
         assert(lng != null),
         assert(zipCode != null);
 
   static MonumentLocation fromJson(Map<String, dynamic> data) =>
-      MonumentLocation(lat: data["lat"], lng: data["lng"], zipCode: data["zipCode"]);
+      MonumentLocation(
+          lat: data["lat"].toDouble(),
+          lng: data["lng"].toDouble(),
+          zipCode: data["zipCode"]);
 }
